Fix unsupported files exporting functions for apparmor and seccomp
